+++ src/main/java/org/jabref/gui/maintable/CellFactory.java
@@ -27,13 +27,9 @@ public class CellFactory {
         // icon.setToo(Localization.lang("Open") + " PDF");
         TABLE_ICONS.put(StandardField.PDF, icon);
 
-        icon = IconTheme.JabRefIcons.DATABASE;
+        icon = IconTheme.JabRefIcons.WWW;
         // icon.setToolTipText(Localization.lang("Open") + " URL");
         TABLE_ICONS.put(StandardField.URL, icon);
-
-        icon = IconTheme.JabRefIcons.DATABASE_PLUS;
-        // icon.setToolTipText(Localization.lang("Open") + " URL");
-        TABLE_ICONS.put(StandardField.URLS, icon);
 
         icon = IconTheme.JabRefIcons.WWW;
         // icon.setToolTipText(Localization.lang("Open") + " CiteSeer URL");
@@ -98,7 +94,6 @@ public class CellFactory {
         icon = printedViewModel.getIcon();
         // icon.setToolTipText(printedViewModel.getLocalization());
         TABLE_ICONS.put(SpecialField.PRINTED, icon);
-
     }
 
     public Node getTableIcon(Field field) {