+++ CHANGELOG.md
@@ -128,6 +128,8 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 - We fixed an issue regarding pasting on Linux. [#6293](https://github.com/JabRef/jabref/issues/6293)
 
 - We fixed linked identifier icon inconsistency[#6705](https://github.com/JabRef/jabref/issues/6705)
+- We fixed the linked identifier icon inconsistency [#6705](https://github.com/JabRef/jabref/issues/6705)
+
 ### Removed
 
 - We removed the option of the "enforce legal key". [#6295](https://github.com/JabRef/jabref/issues/6295)